config structs to file
